requirements.txt
allure-pytest==2.8.21
pytest==6.1.2
pytest-xdist==2.1.0
requests==2.25.0